refactor(entities-plugin): freeform - rename jwt-signer form to common form
