+++ packages/entities/entities-plugins/src/components/free-form/Common/ConfigForm.vue
