+++ packages/entities/entities-plugins/src/components/free-form/Common/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CommonForm.vue'
