--- packages/entities/entities-plugins/src/components/free-form/JwtSigner/index.ts
@@ -1,1 +1,0 @@
-export { default } from './JwtSigner.vue'
