+++ packages/entities/entities-plugins/src/components/free-form/index.ts
@@ -2,4 +2,4 @@ export { default as RequestCalloutForm } from './RequestCallout'
 export { default as ServiceProtectionForm } from './ServiceProtection'
 export { default as DatakitForm } from './Datakit'
 export { default as AIMcpForm } from './AIMcp'
-export { default as JwtSignerForm } from './JwtSigner'
+export { default as CommonForm } from './Common'
